+++ server/odc-service/src/main/java/com/oceanbase/odc/service/exporter/model/ExportConstants.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.service.archiver.model;
+package com.oceanbase.odc.service.exporter.model;
 
 public final class ExportConstants {
