+++ server/odc-service/src/main/java/com/oceanbase/odc/service/exporter/streamprovider/RemoteUriStreamProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.service.archiver.streamprovider;
+package com.oceanbase.odc.service.exporter.streamprovider;
 
 import java.io.InputStream;
 import java.net.HttpURLConnection;
