+++ server/odc-service/src/main/java/com/oceanbase/odc/service/schedule/ScheduleExportFacade.java
@@ -15,7 +15,7 @@
  */
 package com.oceanbase.odc.service.schedule;
 
-import com.oceanbase.odc.service.archiver.model.ExportProperties;
+import com.oceanbase.odc.service.exporter.model.ExportProperties;
 
 public interface ScheduleExportFacade {
