+++ server/odc-service/src/main/java/com/oceanbase/odc/service/schedule/archiverist/ScheduleTaskExporter.java
@@ -15,9 +15,9 @@
  */
 package com.oceanbase.odc.service.schedule.archiverist;
 
-import static com.oceanbase.odc.service.archiver.model.ExportConstants.ARCHIVE_TYPE;
-import static com.oceanbase.odc.service.archiver.model.ExportConstants.FILE_NAME;
-import static com.oceanbase.odc.service.archiver.model.ExportConstants.SCHEDULE_ARCHIVE_TYPE;
+import static com.oceanbase.odc.service.exporter.model.ExportConstants.ARCHIVE_TYPE;
+import static com.oceanbase.odc.service.exporter.model.ExportConstants.FILE_NAME;
+import static com.oceanbase.odc.service.exporter.model.ExportConstants.SCHEDULE_ARCHIVE_TYPE;
 import static com.oceanbase.odc.service.common.util.OdcFileUtil.createFileWithDirectories;
 
 import java.io.File;
@@ -38,12 +38,12 @@ import com.oceanbase.odc.common.security.PasswordUtils;
 import com.oceanbase.odc.common.util.FileZipper;
 import com.oceanbase.odc.metadb.schedule.ScheduleEntity;
 import com.oceanbase.odc.metadb.schedule.ScheduleRepository;
-import com.oceanbase.odc.service.archiver.ExportConfiguration;
-import com.oceanbase.odc.service.archiver.Exporter;
-import com.oceanbase.odc.service.archiver.model.ExportProperties;
-import com.oceanbase.odc.service.archiver.model.ExportRowDataAppender;
-import com.oceanbase.odc.service.archiver.model.ExportRowDataMapper;
-import com.oceanbase.odc.service.archiver.model.ExportedFile;
+import com.oceanbase.odc.service.exporter.ExportConfiguration;
+import com.oceanbase.odc.service.exporter.Exporter;
+import com.oceanbase.odc.service.exporter.model.ExportProperties;
+import com.oceanbase.odc.service.exporter.model.ExportRowDataAppender;
+import com.oceanbase.odc.service.exporter.model.ExportRowDataMapper;
+import com.oceanbase.odc.service.exporter.model.ExportedFile;
 import com.oceanbase.odc.service.connection.ConnectionService;
 import com.oceanbase.odc.service.connection.database.DatabaseService;
 import com.oceanbase.odc.service.connection.database.model.Database;
